fix: 修复 dayShopGoodsQuery 数据类型转换问题
- totalQty 从 int 改为 double 支持小数值
- 使用 Double.parseDouble() 替代 Integer.parseInt()

+++ REPORT/src/main/java/com/report/service/impl/ReportServiceImpl.java
@@ -352,7 +352,7 @@ public class ReportServiceImpl implements ReportService {
 
             // 计算汇总数据
             double totalAmount = 0;
-            int totalQty = 0;
+            double totalQty = 0;
             
             for (Map<String, Object> row : resultList) {
                 Object amountObj = row.get("AMOUNT");
@@ -361,7 +361,7 @@ public class ReportServiceImpl implements ReportService {
                 }
                 Object qtyObj = row.get("QTY");
                 if (qtyObj != null) {
-                    totalQty += Integer.parseInt(qtyObj.toString());
+                    totalQty += Double.parseDouble(qtyObj.toString());
                 }
             }
